Blacklist moved to a new behavioral cog, prune_bot_collections.py Per-user and per-guild blacklisting removed for now.
